Move globals to separate file for use them with require
