+++ js/components/CreateArticleList.js
@@ -1,4 +1,6 @@
-var React = require('react');
+var React = require('react'),
+	CreatArticleLists = require('./CreatArticleLists'),
+	CreatArticleEl = require('./CreatArticleEl');
 
 var CreateArticleList = React.createClass({
 	render: function() {
